Working sump monitor.
This is a working monitor of the sump pump. It connects to the RaspberryPi
that is monitoring the actual sump pump. The data indicates the state of the
pump and the timestamp of the state change.

This monitor implements a state machine that handles the OFF-ON-OFF changes.
It calculates the duty cycle and water inflow rate and then posts the three
pieces of info to a redis database running on localhost.