Don't use the strict TOML solution.
The `toml` crate works with `serde` to serialize/deserialize
TOML documents. I believe, from the minimal documentation, that
all fields must be present (or defined as Option<>) for this to
work. And it doesn't say what happens if a user adds an unknown
key.

So this commit, instead, uses a toml::Value so that users can
safely add any keys without breaking the loading of the config
(we'll simply ignore unknown keys.)