Revert part of commit adbffbfc.
Always sending the OFF event didn't work. The timestamp of the OFF
event is set and doesn't change until the next OFF event. The redis
database won't let you insert timestamps that are the same or less
than the last timestamp so the task was getting an error everytime I
tried restarting the system.