Add redis database to the config.
`redis` supports 16 separate database instances. By default
applications typically use database 0. But, for testing, it would be
nice to use a non-production database.

This commit adds a database number field to the config file as well as
a command line option to specify it. The `redis` connection function
will use this configured value to open the appropriate database,