Fix spelling mistake.
Rather than say "no output" (actually "no ouput"), output the value of
the stream. "no output" was to mean no new value was generated, if
this was implemented as streams. Logic-wise we should show the
previous value -- even if the implementation chooses to optimize out
the fact the value didn't change.