Create methods that return one key.
Originally, `.get_keys()` would return a pair of keys -- one for
device information and the other for time-series data. With settings,
we need yet a third key. Rather than return a 3-tuple (and build all
three keys when some drivers only need one of them), the names of keys
is returned with separate methods.