change: reduce light controls for sump pump
The sump driver has been working well for several months. We don't
necessarily need the lights to cycle when it turns on -- especially
since we use the Philips color lights for holiday decorations.

This change makes the driver only control the lights when an urgent
condition happens with the pump.