refactor: move `driver::Context`.
`driver::Context` is now `device::db::Context`.

A driver interacts with hardware and a `device::Device` type. The
code to interact with `redis` doesn't belong in the `driver` module.
The `device` module, on the other hand, is all about mapping device
interactions with the `redis` API. So, although a driver may receive a
`device::db::Context`, it can only use the public API.