fix: make adjustment based on tests,
The tests showed that `redis::Value::Nil` could be translated to
`data::Type::Nil`. I really only want `redis::Value::Data()` to
be used to store drmem values. This commit enforces this.

(The previous version used `from_redis_value()` to convert to a
`Vec<u8>`. Unfortunately a `redis::Value::Nil` got converted to an
empty vector which we happily conveted to a `data::Type::Nil`.)