refactor: massive project re-organization
Re-organize the project so it uses cargo's "workspaces" feature. This
organizes the various pieces of `drmem` into separate crates, but
keeps the project together as a whole.

This commit results in a buildable version of `drmemd` using the new
organization. However, that doesn't mean the conversion is complete;
the driver, for instance, is still in the `drmemd` binary crate.
Future commits will complete the transistion.

The following subdirectories have been made:

- `drmemd` -- holds the crate that builds the final `drmem`
  executable.
- `drmem-driver-api` -- holds the crate that defines traits and types
  used throughout `drmem` and its drivers.
- `drmem-config` -- holds the configuration logic.
- `backends` -- this subdirectory holds crates that provide storage
  backends for `drmem. Right now it only contains a crate for Redis.
  Eventually I'd like to see other options, like a "simple" backend
  that doesn't provide any more than one point of history. There could
  also be backends that use an SQL-based database.
- `driver` -- this subdirectory will hold crates that implement
  drivers.