refactor: Put sump driver in its own crate.
This is another big refactorization. This commit moves the sump pump
driver into its own crate and preps the entire framework for the
driver API. As I was re-organizing, I realized some of my ideas
weren't going to work out. This commit isn't how I'd like the final
version to be, but I wanted it to be refactored and be able to compile
and run. Future commits will provide the final solution.