fix: improve handling of redis config.
If the `redis-backend` feature is specified, the `drmem-config` crate
will define a structure to be used in the `.toml` file. If no section
is added to the config file, sensible defaults are used. Also, debug
and release versions will set different values for the database
number (redis supports 16 different dataabases -- drmem now uses #0
for operational stuff and #1 for development and testing.)