fix: get config file support working.
This commit is the result of adding tests and fixing the code so it
works as the tests require.

Rather than have a `redis` module and field, which requires a lot of
conditional compilation directives, this commit creates a `backend`
module. DrMem needs one backend storage module and one of several will
be chosen. But all have the same module name so the only conditional
part is which module defines the config section.