feat: add new feature: `graphql`.
This commit removes the GraphQL interface from the default build. If
someone wants to access DrMem through the GraphQl interface, then they
can enable this feature.

As of this commit, however, the `juniper_hyper` crate pulls in an old
version of `hyper` which pulls in an old version of `tokio`. This is
one of the reasons I made it an optional feature.