:triangular_flag_on_post: use "thin" LTO
For releases, we use link-time optimizations (LTO) to remove dead code
from the executable and perform program-wide optimizations. This
normally adds a lot of time to the linking stage. The LLVM toolchain
now has a "thin" option for LTO in which it works almost as well as
"fat" LTO (which was the default), but completes much faster.