:truck: move LICENSE text
Remove copyright from the source files; projects don't seem to put
copyrights in the code but, instead, have a LICENSE file that sets the
copyright for the whole project.

Since this project will be publishing several subprojects, I copied
the LICENSE file into the base directory of each, as well.