:building_construction: massive changes
I typically try to make small commits but this refactoring,
unfortunely, lead to too many overlapping changes and it would have
been a lot of work to break it into separate commits so I'm putting
it all in this commit. The fact that the project didn't build a useful
executable before this commit means the inidividual commits aren't too
important at this point of the project.

Some notable changes:

- removed the `drmem-types` crate and put its contents into
  `drmem-api`.
- added a simple backend crate (`drmem-db-simple`)
- created a `Store` trait which backends implement