:sparkles: add support to load drivers from config
This commit adds support to load drivers specified in the `drmem.toml`
file. The config file has a section where an array of drivers is
selected. The core now iterates across the array and starts instances
of the driver. Each instance is fed config parameters (specific to
each driver.)

The sump pump driver has been updated to work with this initialization
framework.