:recycle: reorganize the sump pump driver
This is a refactoring of the sump pump driver, and fixing the API to
support the new layout (which consisted mostly of adding constraints
to the traits.)

The driver used to retry connecting to the remote task when the TCP
connection broke. But this is what the driver monitors do (restart the
driver after a failure.) So now the driver doesn't need to do as much.