:sparkles: add read-write support for `redis` backend
This commit completes (minus unit tests) `redis` support for the
latest back-end API. The RedisStore type now has a HashMap with a
device-to-mpsc::Sender association so clients can send settings to a
driver.

This commit also adds support for returning to the driver the last
value the device was reported.