:sparkles: registering RO device gets previous value
When registering a read-only device, the prior value can, optionally,
be returned. Most devices will ignore this value and the simple
back-end always returns `None`.

There may be a small set of devices that would like to use the prior
value for initialization. A device integrating a reading, for
instance, may want to start at the previous value instead of at 0.