:sparkles: the simple backend handles re-registrations
If a driver crashes, it gets restarted. The associated devices will,
therefore, get registered again. The REDIS backend accepted multiple
registrations. The simple backend did not.

This commit attaches the driver name to the devices so, when a device
gets re-registered, the backend can verify it's the same driver and
allow it.