:sparkles: the simple backend now reports previous values
When a driver registers a device, it receives a tuple. One of the
elements of the tuple is the last known value of the device. For the
simple backend, this was always set to `None`. But if a driver crashes
and gets restarted, it would be nice to report the previous value.

This commit has the `ReportReading` closure also save a copy of the
reading in the `DeviceInfo` table.