:bug: fix race condition in simple backend
The `broadcast::Sender` has been added to the context protected by the
`Mutex`. The code was correct before, but only due to coincidence. By
moving the handle inside the mutex, the code guarantees the last
reading and the broadcast channel are in sync.

This will become more important when the client API subscribes to
readings.