:zap: improve delay accuracy
The `update_state()` method is very quick to compute and it could
return a new expiration time. So this commit moves this calculation to
the start of the method, before a reply is sent to the client. The
communication channel can take several milliseconds which was pushing
the timeout by tens of milliseconds.