:building_construction: driver API changes
Previously, a driver could return `Ok(())` although, by convection, it
never should. It could also return `Err(_)` to report a problem. In
addition, since drivers are run as spawned tasks, there was the need
to check the return value of the `JoinHandle` to detect panics and
cancellations.

This commit changes the driver API so that drivers now return
`Infallible`, which is to say they can't return anything and,
therefore, run forever.

If a serious error occurs, a driver should call `panic!()`. The
management layer is *much* simpler and only needs to catch and report
the panics (afterwhich, it restarts the driver.)