:recycle: change encoding for booleans
To be efficient, booleans were encoded as `"T"` or `"F"`. But we need
a way to query the type of a device. IRedis, the type will be stored
in the database. We don't, however, want to check for `"T"` or `"F"`
so we use `'B'`.

Also, I envision us adding array types down the road and they're
encode will be an `'A'` followed by the type. It would be annoying to
check for `"AT"` and `"AF"` for an array of booleans.