:recycle: refactor code for tests
We aren't going to write tests for redis since its project team will
be testing it and it would make it harder for GitHub to run our tests.
But we can test that we generate the Redis commands properly.

This commit pulls out the command-making code and places it in their
own, small function. Tests have been added to verify we're generating
the proper commands.