:art: ignore results in the pipeline
For this case, we're initializing a device's records in redis. We're
not interested in any of the results, yet a pipeline will keep
building a larger and larger tuple as its return type. The redis
commands were careful crafted to work in the order they are defined,
so we don't need any of the results. By using `.ignore()`, the
pipeline doesn't extend the size of the tuple result.