:bug: give each GraphQL client a connection to redis
The multiplexed connection in the redis crate is shared with every
async task. When a client is monitoring a device, it is blocked until
the next point arrives. Since the connection can only handle one
command at a time, when a client is blocked, no new device updates are
made!

This commit changes the subscription handler to create a new, unshared
connection to redis for the client's request.