:sparkles: add GraphQL config to `.toml` file
This commits adds an `addr` parameter to a GraphQL configuration
section. If not specified, it defaults to 0.0.0.0:3000.

Had to rearrange things a bit. Now we get the configured address and
have the http server bind to it. The resulting address (which may
change, if an ephemeral port was requested) is then fed to the mDNS
system. So if you want multiple DrMem processes running on a machine,
you can let them chose ephemeral ports and mDNS will still let you
find them.