:sparkles: add a 'name' option for GraphQL clients to use
The TOML file now has a `name` parameter which is used by the mDNS
responder to create an advertized name. Each instance of DrMem should
have a unique name. If a name is not provided, the `HOST` environment
variable is used.

I chose to use the environment variable because `gethostname()` is not
provided by the Rust standard library and I didn't want to bring in
another crate.