:recycle: add type-safety to incoming settings
Previously, drivers would get an `mpsc::Receiver` and would have to do
their own conversions from `device::Value` to a primitive, Rust type.
The reply channel also required the driver to convert to a
`device::Value` when replying.

With this commit, a driver gets a `Stream` of setting requests where
the data has already been converted to the desired type. Any requests
that send the wrong data type will get an error reply so drivers
never have to deal with that error checking.