:sparkles: add lexing and parsing
This commit introduces the lex and yacc facilities in Rust. In
`drmemd`, there is a `logic` module which handles parsing the logic
configuration expressions.

Right now there's a data structure that the parser generates. It
doesn't yet get used, but the tools are in place and a simple grammar
has been defined and works.

I even added a few tests to make sure it was working like I expected.