:bug: remove bad test
Removed the line that was breaking the build because it was an invalid
test. All accesses of the log level should be done through the
"getter" method -- which was thoroughly tested in the global tests.
This was an outdated test that was hiding in the tests for the
`redis-backend` feature.