:sparkles: stop error recovery in parsing
The lex/yacc crate we're using has a powerful recovery system when
they can't parse a grammar. In our case, our expressions are one line
long, instead of a file full of source code. And, being an embedded
system, we can't easily provide the user with all the corrective
recommendations. So this commit disables the error correction and
simply tells the user where in the line there was an error.