:art: fix rust formatter
We're in a situation where my Rust installation is older than the
tools available on GitHub. A parameter for the Rust formatter has
changed and I don't know how to make it work for both versions.
However, I like the way it formats function parameters when I
comment-out the config, so I'll commit this rather large change.