:bug: wasn't handling empty history correctly
For the REDIS backend, if the device had no history, `get_device_info`
was returning an error instead of an empty history indicator. This
made the mobile app spin (since no GraphQL reply was being sent.)

I didn't add a test for this because I'm not sure how to model it in
the test framework.