:zap: improve sending commands
The previous version was using `.send_*` calls which resulted in
little packets being sent. This caused delays in the communication.
This version is a little more verbose and copies the data, but the
entire command (including length header) all gets sent at once and the
device responds better to it.