:zap: streamline writing the request
Previous to this, `serde_json` serialized the command to a vector,
then the vector was encrypted, and, lastly, a new vector was created
with the length header and then the encrypted buffer copied to it.
That's too many copies.

In this version, we create a `CmdWriter`, which implementes the
`Write` trait. As it writes, it encrypts. So now `serde_json` emits
the serialized data ... and we're done.