:zap: maintain local buffer for reads
The `Instance` type now holds a buffer to receive replies. Much of the
`Instance` methods now take a `&mut self`. These methods, also, no
longer take a `TcpStream` argument but, instead, take a `&mut
AsyncReadExt`. This allows us to write tests that don't need to open
sockets (since `&[u8]` views implement the traits.)

Previously, the `read_reply()` method would create a local, 1000-byte
buffer to store the incoming message, then decrypt it, then
deserialize it. With this change, we remove the initialization of a
1000-byte buffer every time we want to receive a reply.