:bug: don't scale "millis" parameter
When the `cycle` driver simply emitted a square wave, the "millis"
parameter indicated the whole cycle so we divided it by 2. Now that it
can be a longer sequence, it just indicates the length of time each
value affects the output.

Added unit tests to make sure this, and other configuration
conditions, are correct.