:recycle: improve `logic::Output::send()`
I never really checked if the previous version was actually comparing
the values correctly. I was creating a `Some(&value)` to do the
comparison with the stored value from the previous setting. I think it
was comparing the references, not the values at the references.

This new version reaches into the `Option` to get the value and
compares with the passed in value. Tests were added to make sure this
was so.

Also, this version makes fewer clones of the value.