:sparkles: add "initial" cfg param to memory driver
This commit adds the "initial" parameter to the driver configuration
for the memory driver. Users can have the memory device take on a
default value when the driver starts up.

This feature was added because the DrMem browser couldn't display a
device's value or know what setting type to use until the device
reported at least one value.