:bug: fix AND and OR short circuiting
The `and` and `or` operators take two booleans. In this case, they
return the expected value. But an expression can also fail (return
`None`) so we needed to augment the behavior. For the `or` operator,
if either operand is true -- even if one is `None`, the result is
true. For the `and` operator, if any input is false, the output is
false -- even if one of the inputs is `None`.