:sparkles: a proposed solution
This took a couple hours to figure out. The compiler kept telling me
things that weren't allowed. The problem was I was trying to pass one
of the channels in the `outputs` array, but the compiler couldn't tell
if more than one future was mutably accessing the same element.

The trick was to pair the output channel with the program so that
iterating through the entries guaranteed each future contained a
unique element.