:wrench: adjust precip total algorithms
Previously, when both the precip rate and precip total went to zero,
the last total would get set to the total before it was reset. It
turns out that many weather stations simply accumulate all the day's
rainfall and resets at night, which is not a very smart algorithm.

The new algorithm is that the total gets reset (and the last total
gets updated) when the rainfall rate is 0 for a certain amount of
time. A Google search found some research that suggested a 10 hour
gap of rain was a good average to indicate separate weather systems.
So after 10 hours of no rainfall, the totals will get updated.