:recycle: encapsulate read-only devices
Up to this point, when a device driver registered a read-only device,
it was given a `Sender<>` handle (to report new values returned by the
device) and the possible previous value of the device. When the driver
API went to the 3-phase lifecycle, the previous value was missplaced
(it couldn't be used during registration.)

With this commit, a new type is defined: `ReadOnlyDevice<T>`. It not
only lets the driver report new values using native types, but it also
hangs on to the previous value. This way the previous value can be
queried during the operational phase of the driver.