:zap: use `Arc<str>` for some constant data
A few fields are sent over the GraphQL channel. When they are `String`
types, they'll get copied. By making them `Arc<str>`, they can never
be altered and only the refernece gets copied on a clone. To make this
work, the "rc" feature needed to be added to the `serde` crate.