:recycle: change how the filters are made
It's difficult to have a function return two different types that are
of the same trait. You end up wrapping it in a pinned box, which then
isn't compatible with the library that wants to use the returned
value.

This commit splits it up in a way that works and also allows us to
write tests to make sure they behave correctly.