:wrench: Remove redis/no-client combo
It doesn't make sense to use the Redis backend without a client API to
access the archived data. This commit removes that combination in the
GitHub workflow (which removes one of the jubs needed for unit tests!)
If the user enables the redis feature, the GraphQL feature also gets
enabled.