:bug: redesign the Future support
Previously, the `Devices` type implemented the `Future` trait
directly. This would work in some cases but, in the case of a
`select!` macro, it would fail because that macro drops all the
futures that didn't resolve. Dropping the future would mean dropping
the Devices type.

In this version, the method `.get_next()` returns a `Future` that
returns the next setting. Since the future isn't directly implemented
by `Devices`, it can safely be "cancelled".