:bug: fix tod and solar start-up
There was a race condition when bringing up the logic blocks. The
`tod` and `solar` tasks will exit when they have no clients. A handle
to their broadcast channel is cloned and given to each logic block as
it starts up. There is a small chance that, before the first logic
block registers for these event, the `tod` or `solar` tasks will check
how many clients they have and see that it's zero and exit.

This commit uses `sync::Barrier` as a synchronization point so the
`tod` and `solar` tasks check their number of clients after every
logic block has initialized.

A Barrier is also passed to each logic block so the for-loop won't
proceed until the currently started logic block has registered for
`tod` or `solar` events.