:art: change param to `&mut`
In the device registration function, the `RequestChan` param was a
move. The problem with this is the driver could save the channel in
the associated type and use it during the life of the driver. The
channel is only intended to be used to register the set of devices.

In this commit, we make the `RequestChan` parameter a mutable borrow
which means it can't save it. (Hmmm. can it make a clone()? Gotta
check.)