:recycle: remove more allocations
The `Store` trait methods no longer return `Pin<Box<Future<>>>`. Now
they just return `impl Future<>`. This reuired a change in the state
of the `core` module since it was storing a `Box<dyn Store>`, which
isn't compatible with `impl Future<>`.

So instead of storing a pointer to a heap-allocated `Store` object, we
now place the `impl Store` directly in the state. This means less use
of the heap and no indirect calls to the methods.