:sparkles: support logic block API for GraphQL
This adds the types and query so GraphQL clients can glean information
about the configured logic blocks in a DrMem node.

This feature isn't complete yet, however. It returns information from
the `Context` of the GraphQL service which, currently, only has an
empty list. The next step is to fill it with config information.