:recycle: slight change to make unit tests work
The change in the logic module broke the unit tests because the
initialization of the logic block happened outside of the async
task. The unit test started an emulator which wasn't ready when the
logic block was making requests.

This change simply clones the Logic config so that we don't have to
make the change to the Logic block initialization.