:construction: partial reconstruction
My house had a power outage in the middle of the night that lasted
longer than my UPS. My SSD's filesystem had some corruption. I was
able to recover some of the work that I did, but lost some, too.

This commit has some of the work and compiles and passes unit tests.
I'll need to recreate the rest of the work in future commits.