:recycle: some shuffling of the types
A lot of the types were too exposed. This moves definitions into areas
that restrict their access. Also got rid of `CmdWriter`; it was a
complication that a lot of areas needed. Instead, I created a `crypt`
module and put simple `encode` and `decode` functions in it.  The
`CmdWriter` was no longer necessary since `Vec<>` is a `Writer` for
the `serde_json` API. :tada: